+++ src/main/java/org/example/plzdrawing/domain/MemberTag.java
@@ -5,6 +5,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.plzdrawing.domain.member.Member;
 
 @Entity
 @Getter
