+++ src/main/java/org/example/plzdrawing/domain/Scrab.java
@@ -6,6 +6,7 @@ import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.plzdrawing.domain.member.Member;
 import org.hibernate.annotations.Where;
 
 @Entity
