+++ src/main/java/org/example/plzdrawing/domain/member/Member.java
@@ -1,31 +1,39 @@
-package org.example.plzdrawing.domain;
+package org.example.plzdrawing.domain.member;
 
 import jakarta.persistence.*;
+import java.util.HashSet;
+import java.util.Set;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
+import org.example.plzdrawing.domain.BaseTimeEntity;
+import org.example.plzdrawing.domain.MemberTag;
+import org.example.plzdrawing.domain.Status;
 import org.hibernate.annotations.Where;
 
-import java.util.ArrayList;
-import java.util.List;
 
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @AllArgsConstructor
 @Where(clause = "status = 'ACTIVE'")
-public class Member extends BaseTimeEntity{
+public class Member extends BaseTimeEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "member_id", nullable = false)
     private Long id;
 
-
     @Column(nullable = false, unique = true)
-    private String username;
-    @Column(nullable = false)
+    private String email;
+
     private String password;
+
+    @Enumerated(EnumType.STRING)
+    @Column(nullable = false)
+    private Provider provider;
+
     @Column(nullable = false)
     private String nickname;
 
@@ -34,17 +42,16 @@ public class Member extends BaseTimeEntity{
     private Status status;
 
     @OneToMany(mappedBy = "member")
-    private List<MemberTag> memberTags = new ArrayList<>();
+    private Set<MemberTag> memberTags;
 
-    public static Member createMember(String username, String password, String nickname) {
-        Member member = new Member();
-        member.username = username;
-        member.password = password;
-        member.nickname = nickname;
-
-        member.status = Status.ACTIVE;
-
-        return member;
+    @Builder
+    private Member(String email, String password, Provider provider, String nickname) {
+        this.email = email;
+        this.password = password;
+        this.provider = provider;
+        this.nickname = nickname;
+        this.status = Status.ACTIVE;
+        this.memberTags = new HashSet<>();
     }
 
     public void addMemberTag(MemberTag memberTag) {
